requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
kafka-python
pypdf
python-dotenv
